Add react and move cesium viewer to a component
