+++ src/components/App.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import CesiumViewer from './CesiumViewer';
+
+const App = () => {
+  return (
+    <CesiumViewer />
+  );
+}
+
+export default App;
