+++ src/css/main.css
@@ -1,4 +1,4 @@
-html, body, #cesiumContainer {
+html, body, #root, .cesiumContainer {
   width: 100%;
   height: 100%;
   margin: 0;
