+++ src/index.html
@@ -2,8 +2,12 @@
   <html>
   <head>
       <meta charset="utf-8">
+      <meta name="viewport" content="width=device-width, initial-scale=1.0">
+      <meta http-equiv="X-UA-Compatible" content="ie=edge">
+      <title>3D Tile Viewer</title>
   </head>
   <body>
     <div id="cesiumContainer"></div>
+    <div id="root"></div>
   </body>
 </html>
